Cocoa: Add support for Vulkan via MoltenVK
This adds basic support for MoltenVK, a Vulkan implementation on top of
Metal, on macOS 10.11 and later.  It looks for MoltenVK in the process
namespace via RTLD_DEFAULT symbol lookup if _GLFW_VULKAN_STATIC is
disabled.

glfwCreateWindowSurface now creates and sets a CAMetalLayer for the
window content view, which is required for MoltenVK to function.

Fixes #870.